requirements.txt
admin==0.0.1
amqp==5.0.2
anaconda==0.0.1.1
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
atomicwrites==1.4.0
attrs==19.3.0
auditlog3==1.0.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.2.1
bulma==1.0.1
cairocffi==1.2.0
CairoSVG==2.5.1
captcha==0.3
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
comtypes==1.1.7
cryptography==3.2.1
cssselect2==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.11
django-admin-tools==0.9.0
django-admin-views==0.8.0
django-allauth==0.43.0
django-betterforms==1.2
django-braces==1.14.0
django-bulma==0.8.1
django-celery-beat==2.1.0
django-channels==0.7.0
django-check-seo==0.3.6
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cleanup==5.1.0
django-cms==3.8.0
django-contactform==2.2.0
django-contactforms==2.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-currentuser==0.5.2
django-db-mailer==2.3.19
django-db-signals==0.1.1
django-debug-toolbar==3.1.1
django-easy-pdf==0.1.1
django-easy-seo==0.4.8
django-extensions==3.1.0
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-heroku==0.3.1
django-hosts==4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-libs==2.0.3
django-log-request-id==1.6.0
django-login-required-middleware==0.5.0
django-mapbox-location-field==1.7.0
django-mathfilters==1.0.0
django-meta==2.0.0
django-model-utils==4.1.1
django-money==1.3
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-payslip==0.3.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-polymorphic==3.0.0
django-pure-pagination==0.3.0
django-push==1.1
django-registration-redux==2.8
django-rest-auth==0.9.5
django-sekizai==2.0.0
django-seo==0.3.5
django-simple-bulma==2.1.0
django-simple-seo==1.0.3
django-social-auth==0.7.28
django-star-ratings==0.9.1
# Editable install with no version control (django-taggit==1.3.0)

# Editable install with no version control (django-taggit-helpers==0.1.4)

django-tags-input==5.0.0
django-timezone-field==4.1.1
django-tinymce==3.1.0
django-treebeard==4.3.1
django-ueditor==1.0.2
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangocms-admin-style==2.0.2
djangocms-page-meta==1.0.1
djangorestframework==3.12.1
DjangoSEO==1.0
docopt==0.6.2
docutils==0.16
easy-thumbnails==2.7.1
entrypoints==0.3
environ==1.0
environs==9.3.0
et-xmlfile==1.0.1
Faker==5.6.3
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
httplib2==0.18.1
idna==2.9
imageio==2.9.0
inflect==5.0.2
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
kiwisolver==1.2.0
kombu==5.0.2
lazy-object-proxy==1.4.3
libsass==0.20.1
lxml==4.6.2
MarkupPy==1.14
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.3.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
num2words==0.5.10
numpy==1.19.2
oauth2==1.9.0.post1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pdfkit==0.6.1
phonenumbers==8.12.11
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.6
py==1.8.1
py-moneyed==0.8.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pyrsistent==0.17.3
pytest==5.4.2
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
PythonTurtle==0.3.2
pyttsx3==2.87
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
reportlab==3.5.59
requests==2.23.0
requests-oauthlib==1.3.0
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sorl-thumbnail==12.7.0
soupsieve==2.1
sqlparse==0.3.1
stripe==2.55.0
swapper==1.1.2.post1
tablib==3.0.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2021.1.14
tinycss2==1.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.56.0
traitlets==5.0.4
twilio==6.41.0
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.25.9
vine==5.0.0
virtualenv==20.0.35
wcwidth==0.1.9
WeasyPrint==52.2
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
